requirements.txt
python-telegram-bot="20.3"
lxml="4.9.2"
requests="2.31.0"
twstock="1.3.1"